+++ python/GafferTest/ValuePlugTest.py
@@ -37,6 +37,10 @@
 
 import gc
 import inspect
+import os
+import subprocess
+import threading
+import time
 import imath
 import six
 
@@ -868,6 +872,104 @@ class ValuePlugTest( GafferTest.TestCase ) :
 
 		with six.assertRaisesRegex( self, BaseException, "Foo" ):
 			GafferTest.parallelGetValue( m["product"], 10000, "testVar" )
+
+	def testCancellationOfSecondGetValueCall( self ) :
+
+		## \todo Should just be checking `tbb.global_control.active_value( max_allowed_parallelism )`
+		# to get the true limit set by `-threads` argument. But IECore's Python
+		# binding of `global_control` doesn't expose that yet.
+		if IECore.hardwareConcurrency() < 3 and "VALUEPLUGTEST_SUBPROCESS" not in os.environ :
+
+			# This test requires at least 3 TBB threads (including the main
+			# thread), because we need the second enqueued BackgroundTask to
+			# start execution before the first one has completed. If we have
+			# insufficient threads then we end up in deadlock, so to avoid this
+			# we relaunch in a subprocess with sufficient threads.
+			#
+			# Note : deadlock only ensues because the first task will never
+			# return without cancellation. This is an artificial situation, not
+			# one that would occur in practical usage of Gaffer itself.
+
+			print( "Running in subprocess due to insufficient TBB threads" )
+
+			try :
+				env = os.environ.copy()
+				env["VALUEPLUGTEST_SUBPROCESS"] = "1"
+				subprocess.check_output(
+					[ "gaffer", "test", "-threads", "3", "GafferTest.ValuePlugTest.testCancellationOfSecondGetValueCall" ],
+					stderr = subprocess.STDOUT,
+					env = env
+				)
+			except subprocess.CalledProcessError as e :
+				self.fail( e.output )
+
+			return
+
+		class InfiniteLoop( Gaffer.ComputeNode ) :
+
+			def __init__( self, name = "InfiniteLoop", cachePolicy = Gaffer.ValuePlug.CachePolicy.Standard ) :
+
+				Gaffer.ComputeNode.__init__( self, name )
+
+				self.computeStartedCondition = threading.Condition()
+				self.__cachePolicy = cachePolicy
+				self["out"] = Gaffer.IntPlug( direction = Gaffer.Plug.Direction.Out )
+
+			# No need to implement `hash()` - because our result is constant (or
+			# non-existent), the default hash is sufficient.
+
+			def compute( self, output, context ) :
+
+				with self.computeStartedCondition :
+					self.computeStartedCondition.notify()
+
+				if output == self["out"] :
+					while True :
+						IECore.Canceller.check( context.canceller() )
+
+			def computeCachePolicy( self, output ) :
+
+				return self.__cachePolicy
+
+		IECore.registerRunTimeTyped( InfiniteLoop )
+
+		for cachePolicy in (
+			Gaffer.ValuePlug.CachePolicy.Legacy,
+			Gaffer.ValuePlug.CachePolicy.Standard,
+			Gaffer.ValuePlug.CachePolicy.TaskIsolation,
+			# Omitting TaskCollaboration, because if our second compute joins as
+			# a worker, there is currently no way we can recall it. See comments
+			# in `LRUCachePolicy.TaskParallel.Handle.acquire`.
+		) :
+
+			node = InfiniteLoop( cachePolicy = cachePolicy )
+
+			# Launch a compute in the background, and wait for it to start.
+
+			with node.computeStartedCondition :
+				backgroundTask1 = Gaffer.ParallelAlgo.callOnBackgroundThread( node["out"], lambda : node["out"].getValue() )
+				node.computeStartedCondition.wait()
+
+			# Launch a second compute in the background, wait for it to start, and
+			# then make sure we can cancel it even though the compute is already in
+			# progress on another thread.
+
+			startedCondition = threading.Condition()
+
+			def getValueExpectingCancellation() :
+
+				with startedCondition :
+					startedCondition.notify()
+
+				with self.assertRaises( IECore.Cancelled ) :
+					node["out"].getValue()
+
+			with startedCondition :
+				backgroundTask2 = Gaffer.ParallelAlgo.callOnBackgroundThread( node["out"], getValueExpectingCancellation )
+				startedCondition.wait()
+
+			backgroundTask2.cancelAndWait()
+			backgroundTask1.cancelAndWait()
 
 	def setUp( self ) :