+++ src/GafferArnold/ArnoldShader.cpp
@@ -184,7 +184,7 @@ namespace {
 	const AtString g_shaderTypeArnoldString( "shaderType" );
 }
 
-static IECore::ConstCompoundDataPtr metadataGetter( const std::string &key, size_t &cost )
+static IECore::ConstCompoundDataPtr metadataGetter( const std::string &key, size_t &cost, const IECore::Canceller *canceller )
 {
 	IECoreArnold::UniverseBlock arnoldUniverse( /* writable = */ false );