+++ src/GafferUIModule/PathListingWidgetBinding.cpp
@@ -256,7 +256,7 @@ class IconColumn : public Column
 		std::string m_prefix;
 		IECore::InternedString m_propertyName;
 
-		static QVariant iconGetter( const std::string &fileName, size_t &cost )
+		static QVariant iconGetter( const std::string &fileName, size_t &cost, const IECore::Canceller *canceller )
 		{
 			const char *s = getenv( "GAFFERUI_IMAGE_PATHS" );
 			IECore::SearchPath sp( s ? s : "" );